임우성 220930_3
14.43

+++ .metadata/.plugins/org.eclipse.wst.server.core/tmp0/wtpwebapps/bookReviewProgram/bkSelectPro.jsp
@@ -9,6 +9,7 @@
 <%@ page import="BKRV.review.rvBean" %>
 <%@ page import="java.util.Vector" %>  
 <%request.setCharacterEncoding("UTF-8");%>
+<%@ page session = "true" %>
 <!DOCTYPE html>
 <html>
 <head>
@@ -25,8 +26,9 @@ margin: auto;
 <body>
 
 <!-- bean이 없는 상태에서 가져올 때는 아래와 같이 이용 -->
-<%
-	int bk_number = Integer.parseInt(request.getParameter("bk_number"));
+<%	String id = (String)session.getAttribute("id");
+	
+    int bk_number = Integer.parseInt(request.getParameter("bk_number"));
 	
 	bkDAO bdao = new bkDAO();
 	bkBean bBean = bdao.selectBook(bk_number);
@@ -117,7 +119,9 @@ margin: auto;
 		</td>
 	</tr>
 	<tr align="center">
-		<td colspan="3"> 				
+		<td colspan="3"> 
+<%	if(id.equals("admin")){	
+   } else {		%>		
 			<button type="button" onclick="location.href='mainSession.jsp?center=bkUpdateform.jsp&bk_number=<%=bk_number%>'">수정</button>&nbsp;&nbsp;
 			<button type="button" onclick="javascript:deleteConfirm()">삭제</button>&nbsp;&nbsp;
 			<script>
@@ -129,6 +133,7 @@ margin: auto;
 					}
 				}
 			</script>
+			<%} %>
 			<button type="button" onclick="javascript:history.go(-1)">이전</button>
 	</tr>	
 </table>
@@ -150,7 +155,7 @@ margin: auto;
 		rvDAO rdao = new rvDAO();
 		Vector<rvBean> rvVec = rdao.allbookselectReview(rv_bknumber);
 
-		String Rv_bknumber = Integer.toString(rv_bknumber);/*  request.getParameter("Rv_id"); */
+		String Rv_bknumber = Integer.toString(rv_bknumber);
 		
 		String column = "Rv_bknumber";
 		String value = Rv_bknumber;
@@ -191,10 +196,12 @@ margin: auto;
 
 <table border="1">
 	<tr align="center">
-		<td>닉네임:&nbsp;<%=rbean.getRv_id() %></td>
+		<td>아이디:&nbsp;<%=rbean.getRv_id() %></td>
 		<td>평점:&nbsp;<%=rbean.getRv_score() %></td>
 		<td>
-		<button type="button" onclick="location.href='mainSession.jsp?center=rvUpdateform.jsp&rv_number=<%=rbean.getRv_number() %>&rv_bknumber=<%=rbean.getRv_bknumber() %>'">수정</button>
+		<%	if(id.equals("admin") || id.equals(rbean.getRv_id())){	
+  		 } else {		%>		
+		<button type="button" onclick="location.href='mainSession.jsp?center=rvUpdateform.jsp&rv_number=<%=rbean.getRv_number() %>&rv_bknumber=<%=rbean.getRv_bknumber() %>&mb_id=<%=(String)session.getAttribute("id")%>'">수정</button>
 		<button type="button" onclick="javascript:rvdeleteConfirm()">삭제</button>&nbsp;&nbsp;
 			<script>
 				function rvdeleteConfirm(){
@@ -208,6 +215,7 @@ margin: auto;
 					}
 				}
 			</script>
+		<%} %>
 		
 		</td>
 	</tr>
@@ -251,12 +259,12 @@ margin: auto;
 		
 		if(startPage > 10){
 	%>
-			<a align="center" href="bkSelectPro.jsp?pageNum=<%=startPage - 10 %>">[previous]</a>
+			<a align="center" href="mainSession.jsp?center=bkSelectPro.jsp&bk_number=<%=bk_number %>&pageNum=<%=startPage - 10 %>">[previous]</a>
 	<%
 		}
 			for(int i = startPage; i <= endPage; i++){
 	%>
-			<a align="center" href="bkSelectPro.jsp?pageNum=<%=i %>">[<%=i %>]</a>
+			<a align="center" href="mainSession.jsp?center=bkSelectPro.jsp&bk_number=<%=bk_number %>&pageNum=<%=i %>">[<%=i %>]</a>
 	<%
 			}
 			
@@ -264,7 +272,7 @@ margin: auto;
 		if(endPage < pagecount){
 			endPage = pagecount;		
 	%>
-			<a align="center" href="bkSelectPro.jsp?pageNum=<%=startPage + 10 %>">[next]</a>
+			<a align="center" href="mainSession.jsp?center=bkSelectPro.jsp&bk_number=<%=bk_number %>&pageNum=<%=startPage + 10 %>">[next]</a>
 	<%
 		}
 	}
